requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlmodel==0.0.22
python-dotenv==1.0.1
twilio==9.3.0
jinja2==3.1.4
httpx==0.27.2
